Cloud network: fix permissions check

+++ infoblox/usecases/impl/CloudNetworkCustomDelete1.py
@@ -1,6 +1,7 @@
 from infoblox.usecases.impl.CloudNetworkDelete import CloudNetworkDelete
 from infoblox.models.Infoblox.Network import Network
 from infoblox.models.History.History import History
+from infoblox.models.Permission.Permission import Permission
 
 from infoblox.helpers.Exception import CustomException
 from infoblox.helpers.Log import Log
@@ -20,16 +21,19 @@ class CloudNetworkCustomDelete1(CloudNetworkDelete):
     # Public methods
     ####################################################################################################################
 
-    def delete(self, *args, **kwargs) -> list:
+    def delete(self, *args, **kwargs) -> None:
         try:
             network = Network(self.assetId, self.networkAddress)
             if not network.repr().get("extattrs", {}).get("Environment", {}).get("value", "") == "Cloud":
                 raise CustomException(status=400, payload={"Infoblox": "This is not a Cloud Network."})
 
-            from infoblox.controllers.CustomController import CustomController
-            network.delete()
-            hid = self.__historyLog(network.network, 'deleted')
-            CustomController.plugins(controller="cloud-network_delete", requestType="network.delete", requestStatus="success", network=network.network, user=self.user, historyId=hid)
+            if Permission.hasUserPermission(groups=self.user["groups"], action="cloud_network_delete", assetId=self.assetId, network=network) or self.user["authDisabled"]:
+                from infoblox.controllers.CustomController import CustomController
+                network.delete()
+                hid = self.__historyLog(network.network, 'deleted')
+                CustomController.plugins(controller="cloud-network_delete", requestType="network.delete", requestStatus="success", network=network.network, user=self.user, historyId=hid)
+            else:
+                raise CustomException(status=403, payload={"Infoblox": "Forbidden."})
 
         except Exception as e:
             raise e
